refactor: `no_std` use `FxHashMap` alias
