+++ cranelift/codegen/src/machinst/abi.rs
@@ -104,10 +104,10 @@ use crate::ir::{ArgumentExtension, ArgumentPurpose, Signature};
 use crate::isa::TargetIsa;
 use crate::settings::ProbestackStrategy;
 use crate::CodegenError;
+use crate::FxHashMap;
 use crate::{ir, isa};
 use crate::{machinst::*, trace};
 use regalloc2::{MachineEnv, PReg, PRegSet};
-use crate::fx::FxHashMap;
 use smallvec::smallvec;
 use std::collections::HashMap;
 use std::marker::PhantomData;
