+++ cranelift/codegen/src/machinst/blockorder.rs
@@ -65,7 +65,7 @@ use crate::entity::SecondaryMap;
 use crate::inst_predicates::visit_block_succs;
 use crate::ir::{Block, Function, Inst, Opcode};
 use crate::{machinst::*, trace};
-use crate::fx::{FxHashMap, FxHashSet};
+use crate::{FxHashMap, FxHashSet};
 
 /// Mapping from CLIF BBs to VCode BBs.
 #[derive(Debug)]
