+++ cranelift/codegen/src/remove_constant_phis.rs
@@ -5,9 +5,9 @@ use crate::ir;
 use crate::ir::Function;
 use crate::ir::{Block, BlockCall, Inst, Value};
 use crate::timing;
+use crate::{FxHashMap, FxHashSet};
 use bumpalo::Bump;
 use cranelift_entity::SecondaryMap;
-use crate::fx::{FxHashMap, FxHashSet};
 use smallvec::SmallVec;
 
 // A note on notation.  For the sake of clarity, this file uses the phrase
